plataforma/requirements.txt
asgiref==3.8.1
authlib ~= 1.0
cffi==1.16.0
cryptography==42.0.8
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.3
django-cors-middleware==1.5.0
django-dotenv==1.4.2
django-grappelli==3.0.8
django-model-utils==4.4.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.1
python-jose==3.3.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
requests==2.26.0  